+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/) and
 the format of [Keep a Changelog](https://keepachangelog.com/).
+
 ## [Unreleased]
 
 ## [5.0.0] - 2025-11-26